+++ src/matcher/err_if.rs
@@ -51,7 +51,7 @@ where
     const CAN_FAIL: bool = true;
 
     #[inline]
-    fn match_with_runner<'a, Runner>(
+    fn match_with_runner<'a, Runner, M: crate::mode::Mode>(
         &'a self,
         runner: &mut Runner,
         error_handler: &mut impl crate::error::error_handler::ErrorHandler,
@@ -62,10 +62,10 @@ where
         'src: 'a,
     {
         let start_pos: usize = input.get_pos().into();
-        match runner.run_match(&self.inner, error_handler, input)? {
+        match runner.run_match::<_, M, _>(&self.inner, error_handler, input)? {
             true => Ok(true),
             false => {
-                if error_handler.is_real() {
+                if M::IS_IN_ERROR_RECOVERY {
                     let ctx = MatchDiagCtx::insertion_point(start_pos);
                     let err =
                         runner.with_snapshot(|snap| self.factory.build_inline_error(ctx, snap));
@@ -74,7 +74,7 @@ where
                         .push_stack_error(ParserError::Inline(err));
                     Ok(true)
                 } else {
-                    Ok(false)
+                    Err(MatcherRunError::RetryRerunNeeded)
                 }
             }
         }
@@ -119,7 +119,7 @@ where
     const CAN_FAIL: bool = Inner::CAN_FAIL;
 
     #[inline]
-    fn match_with_runner<'a, Runner>(
+    fn match_with_runner<'a, Runner, M: crate::mode::Mode>(
         &'a self,
         runner: &mut Runner,
         error_handler: &mut impl crate::error::error_handler::ErrorHandler,
@@ -130,7 +130,7 @@ where
         'src: 'a,
     {
         let start_pos: usize = input.get_pos().into();
-        let matched = runner.run_match(&self.inner, error_handler, input)?;
+        let matched = runner.run_match::<_, M, _>(&self.inner, error_handler, input)?;
         if matched {
             let end_pos: usize = input.get_pos().into();
             let ctx = MatchDiagCtx {