+++ src/parser/multiple.rs
@@ -45,7 +45,7 @@ where
     const CAN_FAIL: bool = Pars::CAN_FAIL;
 
     #[inline]
-    fn parse(
+    fn parse<M: crate::mode::Mode>(
         &self,
         context: &mut ParserContext<'src>,
         error_handler: &mut impl ErrorHandler,
@@ -55,7 +55,7 @@ where
         Inp: Input<'src>,
     {
         let mut results = Vec::new();
-        while let Some(result) = self.parser.parse(context, error_handler, input)? {
+        while let Some(result) = self.parser.parse::<M>(context, error_handler, input)? {
             results.push(result);
         }
         Ok(Some((self.combine_fn)(results)))